+++ src/main/java/com/desafiopicpay/picpay/domain/transaction/TransactionResponseDTO.java
@@ -1,11 +1,12 @@
 package com.desafiopicpay.picpay.domain.transaction;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record TransactionResponseDTO(
         UUID transactionID,
         UUID payerID,
         UUID payeeID,
-        Long value) {
+        BigDecimal value) {
 
 }
