+++ src/main/java/com/desafiopicpay/picpay/domain/users/UsersDTO.java
@@ -1,11 +1,13 @@
 package com.desafiopicpay.picpay.domain.users;
 
+import java.math.BigDecimal;
+
 public record UsersDTO(
         String firstName,
         String lastName,
         String CPF,
         String email,
         String password,
-        Long balance,
+        BigDecimal balance,
         UserType userType) {
 }
